combine_tech_reps() now supports relabeling protein groups
combine_tech_reps() now has a new argument "relabel" that takes a named list of patterns and replacement strings for relabeling/combining protein groups. This is especially useful for handling ambiguous peptides that may belong to multiple protein groups. In the future, it may be better to handle this relabeling on make_auc_table?