Regex to extract protein ID information is now user defined.
To annotate peptides with the proteins that the peptides matched to, I used regular expressions to extract a unique protein ID from the Description column. This column is typically formatted like a fasta header, beginning with a ">" symbol. This formatting depends on how the reference proteome used in the database search was originally formatted. Thus, instead of using a fixed regex, the user can now define the regex with which to extract protein ID information from the description.